+++ wrapper/src/api/providers/benchmark/diagram/diagramAlgorithmSoftKPIsDataProvider.ts
@@ -1,11 +1,9 @@
 import {
   Algorithm,
   DiagramExperimentItem,
-  Experiment,
   SoftKPIsAlgorithmEnum,
 } from '../../../server/types';
 import { AlgorithmProvider } from '../../algorithm/algorithmProvider';
-import { ExperimentProvider } from '../../experiment/experimentProvider';
 import { DiagramDataProvider } from './diagramDataProvider';
 
 export class DiagramAlgorithmSoftKPIsDataProvider extends DiagramDataProvider {
@@ -13,17 +11,15 @@ export class DiagramAlgorithmSoftKPIsDataProvider extends DiagramDataProvider {
     metric: SoftKPIsAlgorithmEnum,
     diagramExperimentItem: DiagramExperimentItem
   ): number {
-    const experimentProvider = new ExperimentProvider();
-    const experiment = experimentProvider.getExperiment(
+    const algorithmProvider = new AlgorithmProvider();
+    const algorithm = algorithmProvider.getAlgorithm(
       diagramExperimentItem.experiment.experimentId
     );
 
-    return this.mapEnum(metric, experiment);
+    return this.mapEnum(metric, algorithm);
   }
 
-  mapEnum(metric: SoftKPIsAlgorithmEnum, experiment: Experiment): number {
-    const algorithmProvider = new AlgorithmProvider();
-    const algorithm = algorithmProvider.getAlgorithm(experiment.algorithmId);
+  mapEnum(metric: SoftKPIsAlgorithmEnum, algorithm: Algorithm): number {
     const mappedMetric = softKPIAlgorithmMap.get(metric);
     if (mappedMetric) {
       if (mappedMetric(algorithm)) mappedMetric(algorithm);
@@ -40,7 +36,7 @@ export class DiagramAlgorithmSoftKPIsDataProvider extends DiagramDataProvider {
       )?.value;
       if (!effort)
         throw new Error(
-          `Either HR-Amount or Expertise is missing for experiment ${experiment.id} so that effort cannot be calculated`
+          `Either HR-Amount or Expertise is missing for experiment ${algorithm.id} so that effort cannot be calculated`
         );
     }
     return effort;
