+++ wrapper/src/api/providers/benchmark/util/enumChecker.ts
@@ -1,4 +1,6 @@
 export function isInEnum<T>(enumToBeChecked: T, value: string): boolean {
+  console.log(value);
+  console.log(Object.values(enumToBeChecked));
   return Object.values(enumToBeChecked)
     .map((enumValue) => enumValue.toString())
     .includes(value);
